Refactor SchedulePage
Rename component and avoid rendering extra ScheduleFinder's
